BOOST REVIVED!
Because there is little documentation for speed and inertia, we can only
guess. I assume all boost does is acceleration, i.e. it should work as
long as we don't throttle. Hard to say; anyway, here it is.

This time there also is a fixed bug that it doens't recognize points —
it seems, points location might change a bit, so consider only its
radius.

Signed-off-by: Konstantin Kharlamov <Hi-Angel@yandex.ru>